backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
google-generativeai
pydantic
scikit-learn
numpy